Wed, Jun 29, 2022, 1:35 AM +00:00

+++ pset4/recover/tst2.c
@@ -11,7 +11,7 @@ typedef uint8_t BYTE;
 
 int main(void)
 {
-    int *buffer;
+    BYTE *buffer;
 
     buffer = (BYTE*)malloc(512*sizeof(BYTE));
